refactor : [FE] Buttons에서 Puzzle 부분 분리
- 퍼즐 부분을 PuzzlePlate로 분리.
- PuzzlePlate를 Puzzle로 또 분리하려 했지만 연관되는 것이 있어서 일단
  놔둠.
- PuzzlePlate는 현재 쓰이는 부분이 아니고 후에 사용될 부분.
